# Lazier return in `UnionType->isSuperTypeOfWithReason()` (~40% faster)